Feature: Fix Create UpdateById

+++ Atos.DevSkills.Domain/IRepository/IBaseRepository.cs
@@ -5,7 +5,7 @@ namespace Atos.DevSkills.Domain.IRepository
     public interface IBaseRepository<T> where T : ModelBase
     {
         Task<List<T>> ListAllActive();
-        Task<T> FindById(long id);
+        Task<T> FindById(int id);
         Task<T> Add(T model);
         Task<T> Update(T model);
         Task Delete(T model);
