+++ Atos.DevSkills.Domain/IService/IDesenvolvedorService.cs
@@ -6,7 +6,7 @@ namespace Atos.DevSkills.Domain.IService
     public interface IDesenvolvedorService
     {
         Task<DesenvolvedorViewModel> CadastrarDesenvolvedorAsync(DesenvolvedorInputModel model);
-        Task<DesenvolvedorViewModel> FindById(long id);
+        Task<DesenvolvedorViewModel> FindById(int id);
         Task<List<DesenvolvedorViewModel>> ListAll();
         Task<List<DesenvolvedorViewModel>> ListAllByskill(string skill);
         Task<DesenvolvedorViewModel> UpdateById(int id, DesenvolvedorUpdateInputModel model);
